refactor: extract number control into a shared component for number and integer
