+++ packages/dataviews/src/dataform-controls/integer.tsx
@@ -1,185 +1,9 @@
 /**
- * External dependencies
- */
-import deepMerge from 'deepmerge';
-
-/**
- * WordPress dependencies
- */
-import {
-	Flex,
-	BaseControl,
-	__experimentalNumberControl as NumberControl,
-	privateApis,
-} from '@wordpress/components';
-import { useCallback, useState } from '@wordpress/element';
-import { __ } from '@wordpress/i18n';
-
-/**
  * Internal dependencies
  */
-import { OPERATOR_BETWEEN } from '../constants';
 import type { DataFormControlProps } from '../types';
-import { unlock } from '../lock-unlock';
+import ValidatedNumber from './utils/validated-number';
 
-const { ValidatedNumberControl } = unlock( privateApis );
-
-type IntegerBetween = [ number | string, number | string ];
-
-function BetweenControls( {
-	value,
-	onChange,
-	hideLabelFromVision,
-}: {
-	value: IntegerBetween;
-	onChange: ( [ min, max ]: IntegerBetween ) => void;
-	hideLabelFromVision?: boolean;
-} ) {
-	const [ min = '', max = '' ] = value;
-
-	const onChangeMin = useCallback(
-		( newValue: string | undefined ) =>
-			onChange( [ Number( newValue ), max ] ),
-		[ onChange, max ]
-	);
-
-	const onChangeMax = useCallback(
-		( newValue: string | undefined ) =>
-			onChange( [ min, Number( newValue ) ] ),
-		[ onChange, min ]
-	);
-
-	return (
-		<BaseControl
-			__nextHasNoMarginBottom
-			help={ __( 'The max. value must be greater than the min. value.' ) }
-		>
-			<Flex direction="row" gap={ 4 }>
-				<NumberControl
-					label={ __( 'Min.' ) }
-					value={ min }
-					max={ max ? Number( max ) - 1 : undefined }
-					onChange={ onChangeMin }
-					__next40pxDefaultSize
-					hideLabelFromVision={ hideLabelFromVision }
-				/>
-				<NumberControl
-					label={ __( 'Max.' ) }
-					value={ max }
-					min={ min ? Number( min ) + 1 : undefined }
-					onChange={ onChangeMax }
-					__next40pxDefaultSize
-					hideLabelFromVision={ hideLabelFromVision }
-				/>
-			</Flex>
-		</BaseControl>
-	);
-}
-
-export default function Integer< Item >( {
-	data,
-	field,
-	onChange,
-	hideLabelFromVision,
-	operator,
-}: DataFormControlProps< Item > ) {
-	const { label, description, getValue, setValue } = field;
-	const value = getValue( { item: data } ) ?? '';
-	const [ customValidity, setCustomValidity ] =
-		useState<
-			React.ComponentProps<
-				typeof ValidatedNumberControl
-			>[ 'customValidity' ]
-		>( undefined );
-
-	const onChangeControl = useCallback(
-		( newValue: string | undefined ) => {
-			onChange(
-				setValue( {
-					item: data,
-					// Do not convert an empty string or undefined to a number,
-					// otherwise there's a mismatch between the UI control (empty)
-					// and the data relied by onChange (0).
-					value: [ '', undefined ].includes( newValue )
-						? undefined
-						: Number( newValue ),
-				} )
-			);
-		},
-		[ data, onChange, setValue ]
-	);
-
-	const onChangeBetweenControls = useCallback(
-		( newValue: IntegerBetween ) => {
-			onChange(
-				setValue( {
-					item: data,
-					value: newValue,
-				} )
-			);
-		},
-		[ data, onChange, setValue ]
-	);
-
-	const onValidateControl = useCallback(
-		( newValue: any ) => {
-			const message = field.isValid?.custom?.(
-				deepMerge(
-					data,
-					setValue( {
-						item: data,
-						value: [ undefined, '', null ].includes( newValue )
-							? undefined
-							: Number( newValue ),
-					} ) as Partial< Item >
-				),
-				field
-			);
-
-			if ( message ) {
-				setCustomValidity( {
-					type: 'invalid',
-					message,
-				} );
-				return;
-			}
-
-			setCustomValidity( undefined );
-		},
-		[ data, field, setValue ]
-	);
-
-	if ( operator === OPERATOR_BETWEEN ) {
-		let valueBetween: IntegerBetween = [ '', '' ];
-		if (
-			Array.isArray( value ) &&
-			value.length === 2 &&
-			value.every(
-				( element ) => typeof element === 'number' || element === ''
-			)
-		) {
-			valueBetween = value as IntegerBetween;
-		}
-		return (
-			<BetweenControls
-				value={ valueBetween }
-				onChange={ onChangeBetweenControls }
-				hideLabelFromVision={ hideLabelFromVision }
-			/>
-		);
-	}
-
-	return (
-		<ValidatedNumberControl
-			required={ !! field.isValid?.required }
-			onValidate={ onValidateControl }
-			customValidity={ customValidity }
-			label={ label }
-			help={ description }
-			value={ value }
-			onChange={ onChangeControl }
-			__next40pxDefaultSize
-			hideLabelFromVision={ hideLabelFromVision }
-		/>
-	);
+export default function Number< Item >( props: DataFormControlProps< Item > ) {
+	return <ValidatedNumber { ...props } decimals={ 0 } />;
 }
